requirements/base.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database (async)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0

# Config and validation
pydantic>=2.0.0
pydantic-settings>=2.0.0


# Auth (for step 3)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File uploads + S3-compatible storage (MinIO)
python-multipart>=0.0.12
boto3>=1.35.0
